funcion eliminar con confirmacion

+++ src/componentes/PedidosChef.jsx
@@ -1,9 +1,9 @@
 import React, { Fragment } from 'react'
 import { db } from '../firebase'
 import '../estilos/pedidoschef.css'
-import Modal from "react-bootstrap/Modal";
 import moment from "moment";
 import 'moment/locale/es';
+import swal from 'sweetalert';
 import growl from 'growl-alert';
 import 'growl-alert/dist/growl-alert.css';
 const effect =
@@ -14,15 +14,12 @@ const effect =
 
 const hmh = require('hmh');
 
-
-
 const PedidosChef = () => {
 
     const [orders, getOrders] = React.useState([])
     const [orderdone, setOrderDone] = React.useState([])
     const [delivery, setDelivery] = React.useState([])
-    const [isOpen, setIsOpen] = React.useState(false);
-    //const [,setDeleteOrder] = React.useState([])
+
 
     React.useEffect(() => {
         db.collection('pedidos').where('status', '==', 'pending').onSnapshot({ includeMetadataChanges: true }, (snap => {
@@ -43,34 +40,24 @@ const PedidosChef = () => {
         )
     }, [])
     
-    const showModal = (id) => {
-     console.log(id,'soy el id que pasa el 1 boton') 
-     setIsOpen(true);
-
-    };
-
-    const deleteOrder = async (id) => {
-        try {
-          await db.collection('pedidos').doc(id).delete()
-          console.log(id) 
-
-        } catch (error) {
-          console.log(error)
-        }
+     const deleteOrder = async (id) => {
+        swal({
+            title: "¿Estas seguro que quieres eliminar este pedido?",
+            icon: "warning",
+            buttons: true,
+            dangerMode: true,
+          })
+          .then((willDelete) => {
+            if (willDelete) {
+              swal("Pedido eliminado con exito!", {
+                icon: "success",
+              });
+               db.collection('pedidos').doc(id).delete()
+            }  else {
+              swal("Your imaginary file is safe!");
+            } 
+          });
     }
-
-    
-    const hideModal = () => {
-      /* console.log(id)  */
-      setIsOpen(false);
-    };
-
-
-    /* const deleteOrder =  => {
-      db.collection("pedidos").doc(order.id).delete()
-      console.log(order.id);
-     }; */
-
 
     const orderDone = (item) => {
         db.collection('pedidos').doc(item.id).update({
@@ -87,52 +74,32 @@ const PedidosChef = () => {
         }
     }
 
-/*          deleteOrder = (item) => {
-        db.collection("pedidos").doc(item.id).delete()
-        .then(() => {
-          console.log("Document successfully deleted!");
-          hideModal()
-      }).catch((error) => {
-          console.error("Error removing document: ", error);
-      });
-    }
- */
     return (
-        <Fragment>
-          
+        <Fragment> 
         <div className="container">
-            <h2>Pedidos a realizar</h2>
+            <p className="caja">Pedidos a realizar</p>
             <div className="row row-cols-3 ">
-                {orders.map((order) => { 
+                {orders.map((order, i) => { 
                 return (
-                    <div className="h5 font-italic">
+                    <div className="h5 font-italic" key={i}>
                         <section className="section" id={order.id}>
                             <div className="row columnLength">
-                            <p className="text client-text"> Dia: {moment(order.orhourDone).subtract(10, 'days').calendar()}</p>
-                                <p className="text client-text"> Hora: {moment(order.orhourDone).format('LTS')}</p>
+                            <p className="text client-text"> Dia: {moment(order.hourSend).format('L')}</p>
+                                <p className="text client-text"> Hora: {moment(order.hourSend).format('LT')}</p>
                                 <p className="text client-text orders"> Cliente: {order.cliente}</p>
                                 <p className="text client-text orders"> Mesas: {order.numMesa}</p>
                                 <span className="menu-name text text-light">Pedido</span>
-                                {order.pedido.map(item => <span className="order-kitchen" key={item.id}>
+                                {order.pedido.map((item, i) => <span className="order-kitchen" key={i}>
                                     <ul>
                                         <li> {item.countProducto} {item.nombreProducto} </li>    
                                     </ul> 
                                 </span>)} 
                                 <div className="orders footer">
-                                    <button  type="submit" className="btn btn-dark" onClick={() => showModal(order)}>Cancelar</button>
+                                    <button  type="submit" className="btn btn-dark" onClick={() =>deleteOrder(order.id)}>Cancelar</button>
                                     <button className="btn btn-light ok" onClick={() => orderDone(order)}>Listo</button>
                                </div>
                             </div>
                         </section>
-                        <Modal show={isOpen} onHide={hideModal}>
-                          
-                                <Modal.Body>¿ Estas seguro que quieres cancelar este pedido ?</Modal.Body>
-                                <Modal.Footer>
-                                <button class="btn btn-dark" onClick={hideModal}>Cancelar</button>
-                                <button  type="submit" class="btn btn-warning"  onClick={() => deleteOrder(order.id)}>Aceptar</button>
-                                 {console.log(order.id)}
-                          </Modal.Footer>
-                          </Modal>
                     </div>
                 )
                 })}
@@ -140,7 +107,7 @@ const PedidosChef = () => {
         </div>
        
         <div className="container">
-          <h2 className="h2">Pedidos entregados</h2>
+          <p className="caja">Pedidos entregados</p>
           <div className="row row-cols-3 ">
               {delivery.map((item, index) => {
               const send = `${new Date(item.hourSend).getHours()}h ${new Date(item.hourSend).getMinutes()}m`;
@@ -152,17 +119,19 @@ const PedidosChef = () => {
                     <section className="section font-italic">
                         <div className="row time">
                             <div className="menu-name">
-                                <p className="text client-text "> Dia: {moment(item.orhourDone).subtract(10, 'days').calendar()}</p>
-                                <p className="text client-text"> Hora: {moment(item.orhourDone).format('LTS')}</p>
+                                <p className="text client-text "> Dia: {moment(item.orhourDone).format('L')}</p>
+                                <p className="text client-text"> Hora: {moment(item.orhourDone).format('LT')}</p>
                                 <p className="card-title orders"> Cliente: {item.cliente}</p>
                                 <p className="card-title orders"> Mesa: {item.numMesa}</p>
-                                <span className="menu-name text text-light">Pedido</span>
+                                <span className="menu-name text">Pedido</span>
                                 {item.pedido.map((item, index) => <span className="order-kitchen" key={index}>
                                  <ul>
                                  <li> {item.countProducto} {item.nombreProducto} </li>    
                                 </ul> 
                                </span>)}
-                                <span className="bg-dark text-light">Tiempo de preparacion:{difftime}</span>
+                               <div className="fondoTime">
+                                <span className="text-light">Tiempo de preparación: { difftime}</span>
+                                </div>
                             </div>
                         </div>
                     </section>
